feat(web-hosting): reduce line for static analysis
ref: #PUWEBPT-33
