+++ packages/manager/apps/web-hosting/src/data/__mocks__/index.ts
@@ -2,4 +2,4 @@ export * from './attachedDomainDigStatus';
 export * from './websites';
 export * from './webHostingDashboard';
 export * from './ssl';
-export * from './managedWordpress';
+export * from './managedWordpress/mocks';
