+++ packages/manager/apps/web-hosting/src/data/__mocks__/managedWordpress/language.ts
@@ -1,0 +1,34 @@
+export const managedWordpressRerefenceAvailableLanguageMock = [
+  {
+    code: 'de_DE',
+    name: 'Deutsch',
+  },
+  {
+    code: 'en_GB',
+    name: 'English (UK)',
+  },
+  {
+    code: 'es_ES',
+    name: 'Español',
+  },
+  {
+    code: 'fr_CA',
+    name: 'Français du Canada',
+  },
+  {
+    code: 'fr_FR',
+    name: 'Français',
+  },
+  {
+    code: 'it_IT',
+    name: 'Italiano',
+  },
+  {
+    code: 'pl_PL',
+    name: 'Polski',
+  },
+  {
+    code: 'pt_PT',
+    name: 'Português',
+  },
+];
