+++ packages/manager/apps/web-hosting/src/data/__mocks__/managedWordpress/mocks.ts
@@ -1,0 +1,4 @@
+export * from './language';
+export * from './ressource';
+export * from './tasks';
+export * from './website';
