+++ packages/manager/apps/web-hosting/src/data/hooks/managedWordpress/managedWordpressReferenceAvailableLanguages/useManagedWordpressReferenceAvailableLanguages.spec.ts
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom';
 import { renderHook, waitFor } from '@testing-library/react';
 import { describe, expect, vi } from 'vitest';
 
-import { managedWordpressRerefenceAvailableLanguageMock } from '@/data/__mocks__';
+import { managedWordpressRerefenceAvailableLanguageMock } from '@/data/__mocks__/managedWordpress/language';
 import { wrapper } from '@/utils/test.provider';
 
 import { useManagedWordpressReferenceAvailableLanguages } from './useManagedWordpressReferenceAvailableLanguages';
