+++ packages/manager/apps/web-hosting/src/data/hooks/managedWordpress/managedWordpressResource/useManagedWordpressResource.spec.ts
@@ -2,23 +2,23 @@ import '@testing-library/jest-dom';
 import { renderHook, waitFor } from '@testing-library/react';
 import { describe, expect, vi } from 'vitest';
 
-import { managedWordpressWebsitesMock } from '@/data/__mocks__';
+import { managedWordpressResourceMock } from '@/data/__mocks__/managedWordpress/ressource';
 import { wrapper } from '@/utils/test.provider';
 
-import { useManagedWordpressWebsites } from './useManagedWordpressWebsites';
+import { useManagedWordpressResource } from './useManagedWordpressResource';
 
-describe('useManagedWordpressWebsites', () => {
+describe('useManagedWordpressResource', () => {
   beforeEach(() => {
     vi.clearAllMocks();
   });
   it('should return resource list', async () => {
-    const { result } = renderHook(() => useManagedWordpressWebsites('test'), {
+    const { result } = renderHook(() => useManagedWordpressResource(), {
       wrapper,
     });
     await waitFor(() => {
       expect(result.current.isSuccess).toBe(true);
     });
 
-    expect(result.current.data).toEqual(managedWordpressWebsitesMock);
+    expect(result.current.data).toEqual(managedWordpressResourceMock);
   });
 });
