+++ packages/manager/apps/web-hosting/src/data/hooks/managedWordpress/managedWordpressResourceDetails/useManagedWordpressResourceDetails.ts
