+++ packages/manager/apps/web-hosting/src/data/hooks/managedWordpress/managedWordpressResourceTasks/useManagedWordpressResourceTasks.spec.ts
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom';
 import { renderHook, waitFor } from '@testing-library/react';
 import { describe, expect, vi } from 'vitest';
 
-import { managedWordpressWebsitesTaskMock } from '@/data/__mocks__';
+import { managedWordpressWebsitesTaskMock } from '@/data/__mocks__/managedWordpress/tasks';
 import { wrapper } from '@/utils/test.provider';
 
 import { useManagedWordpressResourceTasks } from './useManagedWordpressResourceTasks';
