+++ packages/manager/apps/web-hosting/src/data/hooks/managedWordpress/managedWordpressWebsiteDetails/useManagedWordpressWebsiteDetails.spec.ts
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom';
 import { renderHook, waitFor } from '@testing-library/react';
 import { describe, expect, vi } from 'vitest';
 
-import { managedWordpressWebsitesDetailsMock } from '@/data/__mocks__';
+import { managedWordpressWebsitesDetailsMock } from '@/data/__mocks__/managedWordpress/website';
 import { wrapper } from '@/utils/test.provider';
 
 import { useManagedWordpressWebsiteDetails } from './useManagedWordpressWebsiteDetails';
