+++ packages/manager/apps/web-hosting/src/data/hooks/managedWordpress/managedWordpressWebsites/useManagedWordpressWebsites.spec.ts
@@ -2,23 +2,23 @@ import '@testing-library/jest-dom';
 import { renderHook, waitFor } from '@testing-library/react';
 import { describe, expect, vi } from 'vitest';
 
-import { managedWordpressResourceMock } from '@/data/__mocks__';
+import { managedWordpressWebsitesMock } from '@/data/__mocks__/managedWordpress/website';
 import { wrapper } from '@/utils/test.provider';
 
-import { useManagedWordpressResource } from './useManagedWordpressResource';
+import { useManagedWordpressWebsites } from './useManagedWordpressWebsites';
 
-describe('useManagedWordpressResource', () => {
+describe('useManagedWordpressWebsites', () => {
   beforeEach(() => {
     vi.clearAllMocks();
   });
   it('should return resource list', async () => {
-    const { result } = renderHook(() => useManagedWordpressResource(), {
+    const { result } = renderHook(() => useManagedWordpressWebsites('test'), {
       wrapper,
     });
     await waitFor(() => {
       expect(result.current.isSuccess).toBe(true);
     });
 
-    expect(result.current.data).toEqual(managedWordpressResourceMock);
+    expect(result.current.data).toEqual(managedWordpressWebsitesMock);
   });
 });
