+++ packages/manager/apps/web-hosting/src/data/types/product/managedWordpress/cms.ts
@@ -1,0 +1,3 @@
+export enum ManagedWordpressCmsType {
+  WORDPRESS = 'WORDPRESS',
+}
