+++ packages/manager/apps/web-hosting/src/data/types/product/managedWordpress/import.ts
@@ -1,0 +1,14 @@
+import { ManagedWordpressCmsType } from './cms';
+
+export type PostImportPayload = {
+  adminLogin: string;
+  adminPassword: string;
+  adminURL?: string;
+  cms: ManagedWordpressCmsType.WORDPRESS;
+  cmsSpecific?: {
+    wordPress?: {
+      language?: string;
+      url?: string;
+    };
+  };
+};
